Crate open_cdd

Bot OpenCDD

Bot asynchrone

Le bot se repose sur la crate tokio pour bénéficier de la gestion des tâches asynchrones.

Bot orienté par composants

La structure du bot est composée de plusieurs composants.

Chaque composant correspond à un domaine de fonctionnalité.

Pour implémenter une nouvelle fonctionnalité

Pour ajouter une nouvelle fonctionnalité, créez un fichier dans le dossier src/component/components. Dans ce fichier, créez une struct implémentant le trait Component. Rendez le module publique dans le module [component::components]. Enfin, ajoutez le composant dans la fonction [Bot::new].

Prenez exemple sur le composant misc si nécessaire.

Modules

bot

Core de l’application. L’initialisation du bot et la gestion des composants se fait dans ce module.

component_system

Un composant est une partie du bot.

config

util

Fonctions utiles pour l’ensemble du projet

Macros

vdq

Macro pour créer un VecDeque de la même manière que vec!

Traits

ResultLog

Trait à implémenter pour logger les erreurs dans la console.

Functions

main