Module open_cdd::component_system

Un composant est une partie du bot.

Un composant à un domaine défini, tel que là modération, la gestion des tickets ou d’autres taches du genre. Chacun peuvent contenir un set de commandes et d’acquisition événements et réponde à certaines taches de leur domaine. Un composant est sensé s’autogérer mais Mais rien n’empêche la communication entre ces derniers.

Re-exports

pub use event::EventDispatcher;

pub use framework::Framework;

pub use framework::FrameworkConfig;

Modules

command_parser

Parseur de commande

components

Module comportant les composants

data

Module de gestion des données des composants.

event

framework

manager

slash

Générateurs de commandes slash via notre API de commandes.

Structs

Context

The context is a general utility struct provided on event dispatches, which helps with dealing with the current “context” of the event dispatch. The context also acts as a general high-level interface over the associated Shard which received the event, or the low-level http module.

Message

A representation of a message over a guild’s text channel, a group, or a private channel.

Enums

CommandMatch

Retour d’une commande

Event

Event received over a websocket connection

Traits

Component

Trait de base des composants

ComponentExt

Type Definitions

ArcComponent