Module open_cdd::component_system::command_parser

Parseur de commande

Ce module a été conçu pour faciliter la déclaration de composants, de groupes et de commandes.

Il permet de déclarer des commandes, de déclarer des groupes de commandes, des composants, des arguments au commande, de définir leur comportement tel que des arguments requis, leur type, leur valeur par défaut, leur valeur minimum et maximum, les role qui y sont permis, etc.

Il permet aussi l’affichage d’aide automatisé des groupes et des commandes par le biais de composant help.

Le parseur de commande foncitonne de la même manière qu’un parseur de ligne de commande:

Chaque partie de la ligne de commande est séparée par des espaces, sauf les arguments quotés. mot1 mot2 "mot 3"

Les groupes et les commandes sont analysés en mot clé au debut de la commande et de manière recursif : group [...sous_groupes...] command

Les arguments avec des tirets tel que -nom_parametre valeur ou -nom_parametre "valeur avec des espaces" sont considérés comme des paramètres. (TODO : ajouter des flags)

Les arguments sans tiret sont considérés comme des arguments de commande.

Le parseur est là pour aidé a concevoir une commande, mais ne fait pas de traitement.

Exemple

let misc_definition = cmd::Group::new("misc")
    .set_help("Commande diverse, sans catégorie, ou de test")
    .add_command(cmd::Command::new("concat")
        .set_help("Permet de concaténer deux chaînes de caractères")
        .add_argument(cmd::Argument::new("string1")
            .set_help("Première chaîne de caractères à concaténer")
            .set_required(true)
        )
        .add_argument(cmd::Argument::new("string2")
            .set_help("Deuxième chaîne de caractères à concaténer")
            .set_required(true)
        )
)

Utilisation: misc concat -string1 "Hello" -string2 " world"

Modules

matching

Structures de retour d’une commande qui a match avec le parseur

Structs

Argument

Argument de commande

Command

Container

Conteneur de commandes ou de groupes

Group

Iter

Node

Noeud de l’arbre de commandes Un groupe peut contenir des sous groupes et des commandes, stockées dans ces noeuds.

Enums

IterType

ParseError

Erreur de parsing

Type

ValueType

The type of an ApplicationCommandOption.

Traits

Named

Trait d’objet nommé.

Functions

hash

split_shell

Convertit une chaine de caractère en groupe d’arguments