Struct open_cdd::component_system::components::tickets::Tickets

pub struct Tickets {
    data: RwLock<Data<DataTickets>>,
    node: Node,
    archives_folder: PathBuf,
}

Le composant de gestion des tickets

Fields
data: RwLock<Data<DataTickets>>

Données persistantes

node: Node

Configuration des commandes

archives_folder: PathBuf

Dossier de sauvegarde des tickets

Dès que les tickets sont supprimés, ils sont enregistrés dans ce dossier.

Implementations

impl Tickets

pub fn new() -> Self

async fn r_command(
&self,
_: &FrameworkConfig,
_ctx: &Context,
_msg: &Message
) -> CommandMatch

Execute les commandes du composant non slash

async fn r_event(&self, ctx: &Context, evt: &Event) -> Result<(), String>

Dispatch un enevement reçu par le bot

async fn on_interaction(
&self,
ctx: &Context,
interaction: &Interaction
) -> Result<(), String>

Dispatch une intération reçue par le bot

async fn commands(
&self,
ctx: &Context,
guild_id: GuildId,
command: Command,
app_cmd: &ApplicationCommandEmbed<'_>
) -> Result<Option<Message>, Option<String>>

async fn on_app_command(
&self,
ctx: &Context,
app_command: &ApplicationCommandInteraction
) -> Result<(), String>

Dispatch les commandes slash reçu par le bot

async fn on_msg_component(
&self,
ctx: &Context,
msg_component: &MessageComponentInteraction
) -> Result<(), String>

Dispatch les composants de message (boutons, menu déroulant) reçu par le bot

async fn on_ready(&self, ctx: &Context) -> Result<(), String>

Méthode éxécuté sur l’evenement Ready du bot

Met à jour le menu déroulant des catégories

async fn on_ticket_create(
&self,
ctx: &Context,
msg_cmp: &MessageComponentInteraction
) -> Result<()>

Créer un ticket

Lorsque l’on sélectionne le menu déroulant, la méthode créer un nouveau ticket avec les permissions requise et informe l’utilisateur de la création du ticket

async fn on_ticket_close(
&self,
ctx: &Context,
msg_cmp: &MessageComponentInteraction
) -> Result<()>

Fermer un ticket

Lorsque l’on clique sur le bouton de fermeture, la méthode archive puis ferme le ticket

fn get_archive_folder() -> Result<PathBuf, Error>

Retourne le chemin du dossier de stockage des tickets

Le dossier est créé s’il n’existe pas

async fn archive_channel(
ctx: &Context,
channel: ChannelId
) -> Result<(), String>

Archive un ticket

Le nom et l’avatar des utilisateur liés au tickets sont enregistrés ainsi que leurs messages

fn create_close_button<S: ToString>(
&self,
cmps: &mut CreateComponents,
label: S,
style: ButtonStyle
)

Créer un bouton de fermeture de ticket

Ce bouton ets attaché au message d’ouverture du ticket

async fn delete_old_creation_message(&self, ctx: &Context) -> Result<()>

Supprime l’ancien message de création de ticket, celui contenant le menu de sélection de catégorie

async fn set_channel(
&self,
ctx: &Context,
guild_id: GuildId,
id: u64
) -> Message

fn create_select_menu(
&self,
cmps: &mut CreateComponents,
categories: &Vec<CategoryTicket>
)

Créer un menu de sélection de catégorie

async fn update_select_menu(&self, ctx: &Context) -> Result<()>

Met à jour le menu de sélection de catégorie

async fn is_a_ticket(
ctx: &Context,
guild_channel: &GuildChannel
) -> Result<bool, String>

async fn is_ticket_owner(
ctx: &Context,
channel: &GuildChannel,
user_by: UserId
) -> Result<bool, String>

async fn is_staff(
ctx: &Context,
guild_id: GuildId,
user_by: UserId
) -> Result<bool, String>

async fn close_channel(
&self,
ctx: &Context,
guild_id: GuildId,
channel_id: ChannelId
) -> Result<(), String>

async fn add_member(
&self,
ctx: &Context,
guild_id: GuildId,
channel_id: ChannelId,
user_by: UserId,
user: UserId
) -> Result<Message, String>

async fn category_add(
&self,
ctx: &Context,
guild_id: GuildId,
name: String,
desc: Option<String>,
channel_id: u64,
prefix: String
) -> Message

Ajoute une catégorie de ticket

async fn category_remove(&self, ctx: &Context, name: String) -> Message

Supprime une catégorie de ticket

async fn categories_list(&self, ctx: &Context, guild_id: GuildId) -> Message

Liste les catégories de ticket du composant